refactor: Remove usages of Persistence Consent State (#875)
